vinod-rawat/requirements.txt
# requirements.txt
torch==2.3.0+cpu
openai-whisper==20231117
gTTS==2.5.3
pydub==0.25.1
